+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Limits: Enforce token and message limit at lower level (not longer required to check `state.completed` for limit enforcement).
 - Limits: Enforce [custom limits](https://inspect.ai-safety-institute.org.uk/errors-and-limits.html#custom-limit) for samples by raising `SampleLimitExceededError`.
 - Tasks: Optional ability for solvers to [yield scores](https://inspect.ai-safety-institute.org.uk/solvers.html#sec-scoring-in-solvers) for a task.
+- Model API: Log model calls that result in bad request errors.
 
 ## v0.3.58 (16 January 2025)